added extra tests
